Special Orange request; support for SoftBody Goal weights for Lattices. Animating detailed clothes with softbody becomes messy, so now we'll try it this way. :)
It simply uses the W (weight) value, as already available in each Lattice
Point. Only had to make it editable;

- NKey panel
- or press W in editmode

Further there's a minimalistic W button in the softbody Panel!

+++ source/blender/src/editview.c
@@ -1462,7 +1462,7 @@ static void do_lattice_box_select__doSelect(void *userData, BPoint *bp, int x, i
 }
 static void do_lattice_box_select(rcti *rect, int select)
 {
-	struct { rcti *rect; short select, pass, done; } data;
+	struct { rcti *rect; int select, pass, done; } data;
 
 	data.rect = rect;
 	data.select = select;
